Mention the PR creator, as well as merger, if they are different people
@rrees suggested this as being useful for teams where the developer and
the merger are typically different people (for instance, where the
merger is a gatekeeper QA) - you still want to let the developer know that
they need to check their stuff!

Many devs- overwhelmed by excess notifications, and perhaps unknowing
of the available controls in GitHub - have configured email filters
that immediately archive GitHub notifications - @mentioning them is
best chance of getting them to know that this is actually relevant to them.