Comment even on PR updates for oldish merges
The WorthyOfCommentWindow (12 hours) is there to prevent a deluge of
spam comments when you add Prout to a repo for the first time.

This tweaks that behaviour to **always** write a comment on the PR, so long
as we've _previously_ set some label state on it (so shouldn't trigger a
spam deluge on initial run).

Thanks to @sihil for mentioning that the behaviour seemed wrong :)