Don't attempt to scan repos that don't contain `.prout.json` files
Cloning a repo just to find out if it contains a .prout.json file can be
pretty CPU & network intensive - for instance when it's a big repo like
https://github.com/guardian/frontend . Fortunately the GitHub API has an
endpoint which can give a full file list of paths in a given commit's tree:

https://developer.github.com/v3/git/trees/#get-a-tree-recursively

I was concerned about the truncation of the returned tree - thanks to
@izuzak of GitHub Support for indicating that it's unlikely that we'll
encounter this behaviour.

Doing this filtering gives a 90% reduction in the number of repos we
attempt to scan for our Guardian installation of Prout - where team
permissions can sometimes get a little... caca.