Handle PRs merged with a squashed commit
With the recent addition of support for squashed pull requests, there's no
longer a supported way to establish a merged PR's location in a Git commit
graph :(

The GitHub API "head.sha" value on a PR has never given the commit id of
the merge commit - instead it gives the id of the head commit pushed by the
user on the PR - but now that users can squash commits in the GitHub UI,
the head commit pushed by the user is *no longer* merged into the history
of the master branch... it's essentially orphaned, while a very beautiful
squashed commit is instead made part of the history of your master branch.

So, if I hit https://api.github.com/repos/guardian/membership-frontend/pulls/1118,
I get a "head.sha" of 1b74112 for this merged PR - but that commit was
never made a part of the history of the master branch. The squash commit -
15ec41a - was. The only place that sha is visible in the GitHub API
response is in the *deprecated* field of `merge_commit_sha` (which is
supposed to relate to the 'test' merge commit).

For Prout, the bot which works out if your pull request is live, the method
of operation requires comparing an observed Git id (baked into the deployed
webapp) with the commit ids of merged PRs - are those PRs in the history of
the active (deployed) commit or not? It can't do that if there's no
supported way to get a representative sha for a PR, one which is part of
the history of the master branch itself.

Q: Is `merge_commit_sha` still deprecated, or will there be a way to
extract the sha of the squash commit from the GitHub API?

```
$ curl --silent -H "Accept: application/vnd.github.polaris-preview" https://api.github.com/repos/guardian/membership-frontend/pulls/1118 | grep 15ec41ac956252e3a4f76ff8eac5cb1b7754673c
  "merge_commit_sha": "15ec41ac956252e3a4f76ff8eac5cb1b7754673c",
```

https://developer.github.com/changes/2016-04-01-squash-api-preview/
https://developer.github.com/v3/pulls/#get-a-single-pull-request
https://www.theguardian.com/info/developer-blog/2015/feb/03/prout-is-your-pull-request-out
guardian/membership-frontend#1118

https://github.com/rtyley/play-git-hub/commit/bbb9d22543fa9e0d6686d0a59e8baecc2299d3cd