Updates coping with Travis CI API Version 3
Travis CI API Version 3 was launched in April 2017, with a note about
the eventual deprecation of V2:

https://blog.travis-ci.com/2017-04-06-api-v3-is-here

In Travis API v3, there appears to be only one documented method of
obtaining an API token - through their command-line tool! (not through
the Travis admin console, which seems like a way more obvious choice...):

https://developer.travis-ci.org/authentication

...as the token apparently doesn't expire, we can get rid of the code
calling the "/auth/github" endpoint, and just read the Travis tokens from
environment variables instead:

* `PROUT_TRAVIS_TOKEN_OPENSOURCE` <- for public repos & travis-ci.org
* `PROUT_TRAVIS_TOKEN_COMMERCIAL` <- for private repos & travis-ci.com

...the auth changes have also bought a few small changes to the required
HTTP request headers when calling the Travis API.

The API endpoint for requesting a build doesn't seem much changed beyond
that.

https://docs.travis-ci.com/user/triggering-builds

However, the response returned by the API endpoint for getting information
about a build really has changed:

#### API V2
* https://docs.travis-ci.com/api/#builds

`GET /builds/{build.id}` - response included `config` (ie build config
(secure values and ssh key removed)" ... given that Prout changes the
config for a build it was good to see this.

#### API V3
* https://developer.travis-ci.org/resource/build#Build
* https://developer.travis-ci.org/resource/build#find

`GET /build/{build.id}` - note that 's' is gone from `/builds/`, and the
config data has gone missing, while all the build metadata has been moved
up to the top-level, rather than being in a `build` node.


This commit also includes a change from using the GitHub's '*combined*
status' for GitHub commit statuses, to requesting the full list of
statuses for a given commit. This is mainly to make the functional tests
more reliable- the 'combined' status only gives the 'latest' result, but it
transpires that when Travis builds occur only a few seconds apart- like
they do in the functional tests - the statuses can be recorded out of
order, and the wrong Travis build will be considered the 'latest'. We can
get round this if instead we get the full list of statuses, and examine
only the ones that are 'Triggered by Prout'.