Update Prout's test suite to tolerate flicker in GitHub's API
Yesterday Prout's integration test suite was repeatedly failing,
but generally failing on differing tests for each run. In at least
one case, this was because the data returned by GitHub's API was
'flickering' (ie taking a while to converge on a consistent result).

This was causing problems for any test that was using the
`scanShouldNotChangeAnything()` assertion - this assertion is usually
made after a state change has been made (eg simulating a commit being
'seen') and then hitting GitHub's API has confirmed that the
corresponding comment/label-changes are now present. If no further
state changes are made, then with `scanShouldNotChangeAnything()`
we're asserting that no further changes should occur (we're
essentially testing that Prout doesn't go into a terrible spam
loop like it did back in October 2014).

However, if GitHub's API is flickering, we might see that the labels
have been added, then move onto `scanShouldNotChangeAnything()`, but
when `scanShouldNotChangeAnything()` then captures the PR state afresh
(prior to performing a scan), that state, supplied by the GitHub API,
might be *old* and *missing* those new labels that we just got a
moment ago in the prior code! After `scanShouldNotChangeAnything()`
does its scan, it will grab the PR state _again_ - and this time it
will probably have converged to the correct new state with the labels,
but this is a *change* compared to the stale state it captured at the
the start of the method.

To fix this, I've updated the `RepoPR.currentPR()` method - instead
of hitting the GitHUb API once to fetch the PR information, it now
make more than 1 attempt, not returning the information until it's
been consistently identical between two requests 1 second apart.